Formatting correction, and readibility
- Formatting errors corrected under "Under the following terms:"
- changed of to or on last line (just bad typing on my end lol)
- Decided to be more inline with BY-SA, by mentioning "they can modify as long as appropriate credit is given"

In terms of the issue opened, I am alright if people even sold (commercially) this on T-shirts honestly it be an interesting prospect to see a person being their one man walking shop with Dogecoin accepted here ;). So in the case of commercially used I stand by the BY-SA override on these images =).